Set rockets ready for displaying in UI
